backend/requirements.txt
Django==3.2.25
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
django-cors-headers==4.3.1
python-decouple==3.8
Pillow==10.1.0
django-filter==23.3
drf-spectacular==0.26.5
openpyxl==3.1.2
